Crate bewu_util

Structs

AbortJoinHandle

A join handle wrapper that will abort the task when dropped.

AsyncLockFile

AsyncTimedLruCache

An async LRU whose entries are valid for only a certian time period.

StateUpdateRx

The receiver for state updates

StateUpdateStream

A Stream wrapper for a state update receiver

StateUpdateTx

The sender for state updates.

Enums

StateUpdateItem

Either the state if the channel lagged, or an update.

Traits

StateUpdateChannelState

A state that is changed via updates.

Functions

parse_ffmpeg_time

Parse an ffmpeg time, returning the time in seconds.

state_update_channel

Make a bounded channel.